requirements.txt
APScheduler
asn1crypto
attrs
Automat
certifi
cffi
chardet
click
constantly
construct
coreapi
coreschema
cryptography
cssselect
defusedxml
demjson
diff-match-patch
Django
django-cors-headers
django-crispy-forms
django-debug-toolbar
django-filter
django-formtools
django-guardian
django-haystack
django-import-export
django-ranged-response
django-redis
django-reversion
django-simple-captcha
django-tinymce
djangorestframework
djangorestframework-jwt
docopt
et-xmlfile
fake-useragent
Flask
Flask-Login
fonttools
future
gerapy
httplib2
hyperlink
idna
incremental
iso8601
itsdangerous
itypes
jdcal
jieba
Jinja2
jsmin
lxml
Markdown
MarkupSafe
mmh3
mysqlclient
odfpy
openpyxl
parsel
pefile
pika
Pillow
pyasn1
pyasn1-modules
pycparser
pycryptodome
pycurl
PyDispatcher
PyExecJS
PyJWT
pymongo
PyMySQL
pyOpenSSL
pyquery
pyserial
pyspider
pytesseract
python-dateutil
python-ptrace
python-scrapyd-api
pytz
pywin32
PyYAML
queuelib
redis
requests
Scrapy
scrapy-redis
scrapy-splash
scrapyd
selenium
serial
service-identity
six
sqlparse
tablib
tblib
tornado
Twisted
tzlocal
u-msgpack-python
unicodecsv
uritemplate
urllib3
w3lib
Werkzeug
Whoosh
WsgiDAV
xlrd
xlwt
zope.interface
